Initial tmdbsimple adoption

+++ src/static/js/movie_display.js
@@ -1,26 +1,29 @@
-document.addEventListener('DOMContentLoaded', function () {
-    const movies = JSON.parse(document.getElementById('json-data').textContent);
-    console.log('movies:', movies);
-
-    const accountStates = JSON.parse(document.getElementById('account-states').textContent);
-    console.log('accountStates:', accountStates);
+document.addEventListener('DOMContentLoaded', () => {
+    const parsedElements = ['json-data', 'rated-movies', 'favorite-movies', 'watchlist-movies']
+        .map(id => [id, JSON.parse(document.getElementById(id).textContent)]);
+    const parsedData = Object.fromEntries(parsedElements);
+    const { 'json-data': movies, 'rated-movies': ratedMovies, 'favorite-movies': favoriteMovies, 'watchlist-movies': watchlistMovies } = parsedData;
 
     fetchLoggedInStatus()
-        .then(isLoggedIn => displayMovies(movies, isLoggedIn, accountStates))
-        .catch(error => console.error('Error with fetch call:', error));
+        .then(isLoggedIn => displayMovies(movies, isLoggedIn, ratedMovies, favoriteMovies, watchlistMovies))
+        .catch(console.error);
 });
 
-async function fetchLoggedInStatus() {
-    const response = await fetch('http://127.0.0.1:5000/api/logged_in');
-    const data = await response.json();
-    console.log('Logged in:', data);
-    return data.logged_in;
-}
+const fetchJSON = async url => {
+    const response = await fetch(url);
+    if (!response.ok) {
+        throw new Error(`Error with fetch call, HTTP status ${response.status}`);
+    }
+    return await response.json();
+};
 
-function displayMovies(movies, isLoggedIn, accountStates) {
+const fetchLoggedInStatus = () => fetchJSON('http://127.0.0.1:5000/api/logged_in')
+    .then(data => data.logged_in);
+
+function displayMovies(movies, isLoggedIn, ratedMovies, favoriteMovies, watchlistMovies) {
     if (Array.isArray(movies)) {
         movies.forEach((movie) => {
-            const movieCard = createMovieCard(movie, isLoggedIn, accountStates);
+            const movieCard = createMovieCard(movie, isLoggedIn, ratedMovies, favoriteMovies, watchlistMovies);
             document.getElementById('movie-list').append(movieCard);
         });
     } else {
@@ -28,8 +31,7 @@ function displayMovies(movies, isLoggedIn, accountStates) {
     }
 }
 
-
-function createMovieCard(movie, isLoggedIn, accountStates) {
+function createMovieCard(movie, isLoggedIn, ratedMovies, favoriteMovies, watchlistMovies) {
     const movieCard = document.createElement('div');
     movieCard.classList.add('movie-card');
 
@@ -54,57 +56,56 @@ function createMovieCard(movie, isLoggedIn, accountStates) {
     movieDescription.classList.add('movie-description');
     movieDescription.textContent = movie.overview;
 
+    const cardButtons = document.createElement('div');
+    cardButtons.classList.add('card-buttons');
+
     const similarMoviesButton = document.createElement('button');
     similarMoviesButton.classList.add('similar-movies-button');
     similarMoviesButton.textContent = 'Similar Movies';
     similarMoviesButton.addEventListener('click', similarMovieRedirect.bind(null, movie.id, movie.title));
+    cardButtons.append(similarMoviesButton);
 
     movieInfo.append(movieName, movieStats, movieDescription);
-    movieCard.append(moviePoster, movieInfo, similarMoviesButton);
+    movieCard.append(moviePoster, movieInfo);
 
     movieCard.dataset.movieId = movie.id;  // Add movie ID as a data attribute for later reference
     movieCard.id = 'movie-card-' + movie.id;  // Add unique ID to each movie card
 
     // If the user is logged in, create and append the Rate Movie button
     if (isLoggedIn) {
-        const userRating = accountStates.find(ratedMovie => ratedMovie.id === movie.id)?.rated.value;
-        const watchlist = accountStates.find(watchlistMovie => watchlistMovie.id === movie.id)?.watchlist;
+        const userRating = ratedMovies.find(ratedMovie => ratedMovie.id === movie.id)?.rating;
+        const watchlist = watchlistMovies.find(watchlistMovie => watchlistMovie.id === movie.id);
+        const favorite = favoriteMovies.find(favoriteMovie => favoriteMovie.id === movie.id);
 
-        const inWatchlist = watchlist !== false;
-        const starContainer = createStarContainer(movie, userRating, movieCard, window.location.pathname);
-       
-        
-        
-        displayWatchlistButton(movie.id, movieCard, inWatchlist);
+        const inWatchlist = watchlist !== undefined;
+        const inFavorites = favorite !== undefined;
+        const starContainer = createStarContainer(movie, userRating, movieCard, window.location.pathname, cardButtons);
 
         if (userRating) {
             movieInfo.append(starContainer);
-            displayDeleteRatingButton(movie.id, movieCard, window.location.pathname);
-        }
-        else {
-            // On other pages, always append the stars
+            displayDeleteRatingButton(movie.id, movieCard, window.location.pathname, cardButtons, userRating);
+        } else {
             movieInfo.append(starContainer);
         }
+
+        displayWatchlistButton(movie.id, cardButtons, inWatchlist);
+        displayFavoriteButton(movie.id, cardButtons, inFavorites);
+        movieInfo.append(cardButtons)
     }
 
     return movieCard;
 }
 
-function createStarContainer(movieData, userRating, movieCard, pathname) {
+function createStarContainer(movieData, userRating, movieCard, pathname, cardButtons) {
     const starContainer = document.createElement('div');
     starContainer.classList.add('stars');
 
     for (let i = 0; i < 5; i++) {
         const star = document.createElement('span');
         star.classList.add('star');
-        // 2, 4, 6, 8, 10
-        // Subtracting 12 because the stars are in reverse order
         star.dataset.value = 12 - (i + 1) * 2;
-        // Unicode character for a star
         star.innerHTML = '&#9733;';
 
-        // If the user has rated the movie and the rating is greater than 
-        // or equal to the star's value, color the star gold
         if (userRating && userRating >= (star.dataset.value)) {
             star.style.color = "gold";
         }
@@ -112,16 +113,15 @@ function createStarContainer(movieData, userRating, movieCard, pathname) {
         starContainer.appendChild(star);
 
         star.addEventListener('click', function () {
-            rateMovie(this, movieData, movieCard, pathname);
+            rateMovie(this, movieData, movieCard, pathname, cardButtons);
         });
     }
 
     return starContainer;
 }
 
-function rateMovie(event, movieData, movieCard, pathname) {
+function rateMovie(event, movieData, movieCard, pathname, cardButtons) {
     const clickedStar = event;
-    console.info('Clicked star:', clickedStar)
     const stars = Array.from(clickedStar.parentElement.children).reverse();
     const clickedStarIndex = stars.indexOf(clickedStar);
 
@@ -133,9 +133,8 @@ function rateMovie(event, movieData, movieCard, pathname) {
         }
     });
 
-    // Send the rating to the server
     const movieId = movieData.id;
-    const rating = (clickedStarIndex + 1) * 2;  // Multiply by 2 because ratings are from 0.0 to 10.0 in increments of 0.5
+    const rating = (clickedStarIndex + 1) * 2;
 
     fetch('http://127.0.0.1:5000/rate_movie/', {
         method: 'POST',
@@ -151,14 +150,13 @@ function rateMovie(event, movieData, movieCard, pathname) {
         .then(data => {
             if (data.success) {
                 console.info('Successfully rated movie:', movieId);
-                displayDeleteRatingButton(movieId, movieCard, pathname)
+                displayDeleteRatingButton(movieId, movieCard, pathname, cardButtons)
             } else {
                 console.error('Error rating movie:', data.error);
             }
         })
-        .catch(error => console.error('Error with fetch call:', error));
+        .catch(console.error);
 }
-
 
 function formatDate(movie_release_date) {
     let date = new Date(movie_release_date);
@@ -185,21 +183,23 @@ function formatDate(movie_release_date) {
     return formattedDate;
 }
 
-function displayDeleteRatingButton(movieId, movieCard, pathname) {
-    const deleteRatingButton = document.createElement('button');
-    deleteRatingButton.classList.add('delete-rating-button');
-    deleteRatingButton.textContent = 'Delete Rating';
-    movieCard.append(deleteRatingButton);
-
-    deleteRatingButton.addEventListener('click', function () {
-        deleteRating(movieId, movieCard);
-        if (pathname === '/rated-movies') {
-            movieCard.remove();
-        }
-    });
-
-    return deleteRatingButton;
+function displayDeleteRatingButton(movieId, movieCard, pathname, cardButtons) {
+    if (!cardButtons.querySelector('.delete-rating-button')) {
+        const deleteRatingButton = createButton('delete-rating-button', 'Delete Rating', function () {
+            deleteRating(movieId, movieCard);
+            if (pathname === '/rated-movies') {
+                movieCard.remove();
+            }
+        });
+        cardButtons.append(deleteRatingButton);
+    }
 }
+
+const getSessionId = () => fetchJSON('http://127.0.0.1:5000/api/session_id')
+    .then(data => data.session_id || Promise.reject('Error fetching session ID: session ID is empty'));
+
+const getAuth = () => fetchJSON('http://127.0.0.1:5000/api/token')
+    .then(data => data.access_token || Promise.reject('Error fetching access token: access token is empty'));
 
 async function deleteRating(movieId, movieCard) {
     try {
@@ -224,38 +224,13 @@ async function deleteRating(movieId, movieCard) {
     }
 }
 
-
-async function getSessionId() {
-    const response = await fetch('http://127.0.0.1:5000/api/session_id');
-    const data = await response.json();
-    const sessionId = data.session_id;
-
-    if (sessionId) {
-        return sessionId;
-    } else {
-        console.error('Error fetching session ID:', "session ID is empty");
-    }
-}
-
-async function getAuth() {
-    const response = await fetch('http://127.0.0.1:5000/api/token');
-    const data = await response.json();
-    const accessToken = data.access_token;
-
-    if (accessToken) {
-        return accessToken;
-    } else {
-        console.error('Error fetching access token:', "access token is empty");
-    }
-}
-
 function resetStarColor(movieCard) {
     const stars = movieCard.querySelectorAll('.star');
     stars.forEach(star => star.style.color = "gray");
 }
 
-function removeDeleteRatingButton(movieCard) {
-    const deleteRatingButton = movieCard.querySelector('.delete-rating-button');
+function removeDeleteRatingButton(cardButtons) {
+    const deleteRatingButton = cardButtons.querySelector('.delete-rating-button');
     deleteRatingButton.remove();
 }
 
@@ -263,58 +238,27 @@ function similarMovieRedirect(movieId, movieTitle) {
     window.location.href = `/similar-movies/${movieId}/?title=${movieTitle}`;
 }
 
-function displayWatchlistButton(movieId, movieCard, inWatchlist) {
-    const watchlistButton = document.createElement('button');
-    watchlistButton.classList.add('watchlist-button');
-
-    // Store the inWatchlist state on the button itself
-    watchlistButton.dataset.inWatchlist = inWatchlist ? 'true' : 'false';
-
-    if (inWatchlist) {
-        watchlistButton.textContent = 'Remove from Watchlist';
-    } else {
-        watchlistButton.textContent = 'Add to Watchlist';
-    }
-
-    watchlistButton.addEventListener('click', function () {
+function displayWatchlistButton(movieId, cardButtons, inWatchlist) {
+    const watchlistButton = createButton('watchlist-button', inWatchlist ? 'Drop from Watchlist' : 'Add to Watchlist', function () {
         modifyWatchlist(movieId, this);
-    })
+    });
 
-    movieCard.append(watchlistButton);
+    watchlistButton.dataset.inWatchlist = inWatchlist ? 'true' : 'false';
+    cardButtons.append(watchlistButton);
 
     return watchlistButton;
 }
 
 async function modifyWatchlist(movieId, watchlistButton) {
-    // Get the inWatchlist state directly from the button
     let inWatchlist = watchlistButton.dataset.inWatchlist === 'true';
 
-    // Toggle the inWatchlist state
     inWatchlist = !inWatchlist;
-
-    // Update the inWatchlist state on the button
     watchlistButton.dataset.inWatchlist = inWatchlist ? 'true' : 'false';
+    watchlistButton.textContent = inWatchlist ? 'Drop from Watchlist' : 'Add to Watchlist';
 
-    let watchlistContents = inWatchlist;
+    let payload = {movie_id: movieId, watchlist: inWatchlist}
 
-    if (inWatchlist) {
-        watchlistButton.textContent = 'Remove from Watchlist';
-    }
-    else {
-        watchlistButton.textContent = 'Add to Watchlist';
-    }
-
-    fetch('http://127.0.0.1:5000/watchlist_movie/', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json'
-        },
-        body: JSON.stringify({
-            movie_id: movieId,
-            watchlist: watchlistContents
-        })
-    })
-        .then(response => response.json())
+    flask_post_request("watchlist_movie/", payload)
         .then(data => {
             if (data.success) {
                 console.info('Successfully watchlisted movie:', movieId);
@@ -322,6 +266,52 @@ async function modifyWatchlist(movieId, watchlistButton) {
                 console.error('Error watchlisting movie:', data.error);
             }
         })
-        .catch(error => console.error('Error with fetch call:', error));
+        .catch(console.error);
 }
 
+function displayFavoriteButton(movieId, cardButtons, inFavorites) {
+    const favoritesButton = createButton('favorite-button', inFavorites ? 'Unfavorite' : 'Favorite', function () {
+        modifyFavorites(movieId, this);
+    });
+
+    favoritesButton.dataset.inFavorites = inFavorites ? 'true' : 'false';
+    cardButtons.append(favoritesButton);
+
+    return favoritesButton;
+}
+
+async function modifyFavorites(movieId, favoritesButton) {
+    let inFavorites = favoritesButton.dataset.inFavorites === 'true';
+
+    inFavorites = !inFavorites;
+    favoritesButton.dataset.inFavorites = inFavorites ? 'true' : 'false';
+    favoritesButton.textContent = inFavorites ? 'Unfavorite' : 'Favorite';
+
+    let payload = {movie_id: movieId, favorite: inFavorites}
+
+    flask_post_request("favorite_movie/", payload)
+        .then(data => {
+            if (data.success) {
+                console.info('Successfully favorited movie:', movieId);
+            } else {
+                console.error('Error favoriting movie:', data.error);
+            }
+        })
+        .catch(console.error);
+}
+
+const flask_post_request = (endpoint, payload) => fetch(`http://127.0.0.1:5000/${endpoint}`, {
+    method: 'POST',
+    headers: {
+        'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(payload)
+}).then(res => res.json());
+
+function createButton(className, textContent, eventListener) {
+    const button = document.createElement('button');
+    button.classList.add(className);
+    button.textContent = textContent;
+    button.addEventListener('click', eventListener);
+    return button;
+}
